Moved stocktwits tests to own module
